requirements.txt
numpy==2.3.3
pandas==2.3.3
matplotlib==3.10.7
pyyaml==6.0.2
